refactor/payment: refactor controller layer
