+++ payment-service/src/main/java/com/artsem/api/paymentservice/controller/api/BalanceApi.java
@@ -1,6 +1,6 @@
 package com.artsem.api.paymentservice.controller.api;
 
-import com.artsem.api.paymentservice.model.dto.IsBalancePositiveDto;
+import com.artsem.api.paymentservice.model.dto.request.AmountRequestDto;
 import com.artsem.api.paymentservice.model.dto.request.InitBalanceRequestDto;
 import com.artsem.api.paymentservice.model.dto.response.BalanceResponseDto;
 import io.swagger.v3.oas.annotations.Operation;
@@ -20,12 +20,7 @@ public interface BalanceApi {
     @Operation(summary = "Charge a balance", description = "Charge a user’s balance with a specified amount")
     ResponseEntity<BalanceResponseDto> charge(
             @Parameter(description = "ID of the balance to charge") Long id,
-            @Parameter(description = "Amount to charge the balance") Long amount
-    );
-
-    @Operation(summary = "Check if the balance is positive", description = "Check whether the user’s balance is positive")
-    ResponseEntity<IsBalancePositiveDto> isBalancePositive(
-            @Parameter(description = "ID of the user to check balance") Long userId
+            @Parameter(description = "Amount to charge the balance") AmountRequestDto amount
     );
 
     @Operation(summary = "Get balance by ID", description = "Retrieve the balance details of a user by their balance ID")
